Rename dashboard.jsx to index.jsx
Update dashboard.jsx to index.jsx and relative routes importing dashboard.
